+++ lib/grape/middleware/error.rb
@@ -45,14 +45,15 @@ module Grape
         handler = options[:rescue_handlers].find(-> { [] }) { |error, _| klass <= error }[1]
         handler ||= options[:base_only_rescue_handlers][klass]
         handler ||= options[:all_rescue_handler]
-        with_option = options[:rescue_options][:with]
-        if with_option.instance_of?(Symbol)
-          if respond_to?(with_option)
-            handler ||= self.class.instance_method(with_option).bind(self)
+
+        if handler.instance_of?(Symbol)
+          if respond_to?(handler)
+            handler = self.class.instance_method(handler).bind(self)
           else
-            fail NoMethodError, "undefined method `#{with_option}'"
+            fail NoMethodError, "undefined method `#{handler}'"
           end
         end
+
         handler
       end